+++ src/gui/options/tabbox.ts
@@ -4,13 +4,6 @@ import { ModEntry } from '../../types'
 export {}
 declare global {
     namespace modmanager.gui {
-        namespace OptionsTabBox {
-            type GuiOption =
-                | sc.OptionInfoBox
-                | modmanager.gui.OptionsOptionRow
-                | modmanager.gui.OptionsOptionButton
-                | modmanager.gui.OptionsOptionInputField
-        }
         interface OptionsTabBox extends ig.GuiElementBase, sc.Model.Observer {
             gfx: ig.Image
 
@@ -24,7 +17,7 @@ declare global {
             tabs: Record<string, modmanager.gui.OptionsTabBox.TabButton>
             tabArray: modmanager.gui.OptionsTabBox.TabButton[]
             tabGroup: sc.ButtonGroup
-            rows: modmanager.gui.OptionsTabBox.GuiOption[]
+            rows: modmanager.gui.OptionsOptionRow[]
             rowButtonGroup: sc.RowButtonGroup
             tabContent: {
                 buttonGroup: Nullable<modmanager.gui.OptionsTabBox['rowButtonGroup']>
@@ -264,17 +257,7 @@ modmanager.gui.OptionsTabBox = ig.GuiElementBase.extend({
                     continue
                 }
 
-                let optionGui: modmanager.gui.OptionsTabBox.GuiOption
-
-                if (option.type == 'INFO') {
-                    optionGui = new modmanager.gui.OptionsOptionInfoBox(option, 431)
-                } else if (option.type == 'BUTTON') {
-                    optionGui = new modmanager.gui.OptionsOptionButton(option, optionI, this.rowButtonGroup, 431)
-                } else if (option.type == 'INPUT_FIELD') {
-                    optionGui = new modmanager.gui.OptionsOptionInputField(option, optionI, this.rowButtonGroup, 431)
-                } else {
-                    optionGui = new modmanager.gui.OptionsOptionRow(option, optionI, this.rowButtonGroup, 431)
-                }
+                const optionGui = new modmanager.gui.OptionsOptionRow(option, optionI, this.rowButtonGroup, 431)
                 this.rows[optionI] = optionGui
                 optionI++